Update PollutionData objects
